corrected the quest entity and quest repository + better file org
