--- backend/src/main/java/cloudflight/integra/backend/quests/model/QuestDto.java
@@ -1,5 +1,0 @@
-package cloudflight.integra.backend.quests.model;
-
-import cloudflight.integra.backend.quests.model.Quest.Difficulty;
-
-public record QuestDto (Long id, String title, String description, Difficulty difficulty, int exp, Long cityId) {}
