+++ backend/src/main/java/cloudflight/integra/backend/quests/repository/QuestRepository.java
@@ -1,0 +1,10 @@
+package cloudflight.integra.backend.quests.repository;
+
+import org.springframework.stereotype.Repository;
+import cloudflight.integra.backend.quests.model.entity.Quest;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+@Repository
+public interface QuestRepository extends JpaRepository<Quest, Long>{
+
+}
