+++ backend/src/main/java/cloudflight/integra/backend/utils/enums.java
@@ -1,0 +1,7 @@
+package cloudflight.integra.backend.utils;
+
+public class enums {
+    public enum Difficulty {
+        Easy, Medium, Hard
+    }
+}
